# frontend/src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Settly Core API",
    "version": "1.0.0",
    "description": "Authoritative HTTP API contract for the Settly real estate platform. Implements RFC 9457 error taxonomy, strict validation, and bilingual domain entities."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "settly_session",
        "description": "Better Auth HTTP-only secure session cookie"
      }
    },
    "schemas": {
...
      "UnauthenticatedProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/unauthenticated"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  401
                ]
              }
            }
          }
        ]
      },
...
      "PropertyListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "totalCount": {
            "type": "integer"
          }
        },
        "required": [
          "items"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/properties": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List published properties",
        "description": "Returns paginated list of published and reserved properties for public browsing.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of properties",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog",
          "Agent"
        ],
        "summary": "Create a draft property listing (P1)",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProperty"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft property created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenProblem"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/properties/{id}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Get property details by ID",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Property record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Property not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundProblem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Catalog",
          "Agent"
        ],
        "summary": "Update a property listing with two-tier moderation (P6)",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProperty"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenProblem"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict / Terminal state",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictProblem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalog",
          "Agent"
        ],
        "summary": "Hard-delete a never-published DRAFT listing (Decision #42)",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Draft listing deleted"
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenProblem"
                }
              }
            }
          },
          "409": {
            "description": "Cannot hard-delete published listing",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/properties/slug/{slug}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Get property details by SEO slug",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Property record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Property not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/properties/{id}/submit": {
      "post": {
        "tags": [
          "Catalog",
          "Agent"
        ],
        "summary": "Submit draft listing for publication review (P2)",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Property submitted for review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden or agent not verified",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenProblem"
                }
              }
            }
          },
          "409": {
            "description": "Invalid transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictProblem"
                }
              }
            }
          },
          "422": {
            "description": "Missing required fields or <3 images",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/properties/{id}/resubmit": {
      "post": {
        "tags": [
          "Catalog",
          "Agent"
        ],
        "summary": "Resubmit rejected listing for review (P5)",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Property resubmitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "409": {
            "description": "Invalid transition",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/properties/{id}/archive": {
      "post": {
        "tags": [
          "Catalog",
          "Agent"
        ],
        "summary": "Archive a published listing (P7)",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Property archived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "409": {
            "description": "Active reservation prevents archiving",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictProblem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/properties/{id}/mark-sold": {
      "post": {
        "tags": [
          "Catalog",
          "Agent"
        ],
        "summary": "Confirm sale of reserved listing (P9)",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Property marked as SOLD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}